Update workflow files, modify package version, and add summary comments
Updated the directory for package manifests in `dependabot.yml` and modified the workflow and job names in `build_and_test.yml`. Removed the pull request trigger and "Pack" step from the build job. In `nuget-publish.yml`, the default package version was updated and the build command was altered. Summary comments were added to several interfaces and their methods. The import statement in `INotificationSender.cs` was removed. Updated the comment in `SendEventNotificationToClientsHandler.cs` and added summary comments to `SignalRSettings.cs`.