Update packages, target frameworks, and Swagger settings
Updated the `Roslynator.Analyzers` package version and several package references in `Infrastructure.csproj`. The target framework for all projects and migrator projects was updated to `net8.0`. Some code in `Startup.cs` and `ConnectionStringSecurer.cs` was commented out. The `UseSwaggerUi3` method was replaced with `UseSwaggerUi`. The title, description, and `SwaggerSettings` in `openapi.json` were updated to reflect the .NET 8 upgrade.